Migrated SPLIT to a separate crate.
