+++ crates/tools/src/tools/mod.rs
@@ -1,3 +1,2 @@
-pub mod split;
 pub mod touch;
 pub mod ts;
